[Move `@kbn/config-schema` to server] `dashboard`
