+++ src/plugins/dashboard/server/config.ts
